correção ao criar config.js

+++ modules/mac/rotas/instalacao.js
@@ -23,20 +23,11 @@ module.exports = app => {
      */
     app.get('/mac/instalacao', async (req, res) =>  {
         const retorno   = {}
-        const Cache     = require(CORE + '/lib/cache.js')
         const ambiente  = require(ROOT + '/config/ambiente')
-        const config    = require(ROOT + '/config/config')
+        const config    = require(ROOT + '/config/config_default')
         const configEnv = require(ROOT + '/config/config_' + ambiente.ambiente)
         const driver    = !!!configEnv.data_sources['default'].driver ? 'mysql' : configEnv.data_sources['default'].driver.toLowerCase()
         const configDb  = configEnv.data_sources['default']
-        const conexao   = await require(CORE + '/orm/database/driver/'+driver)(configDb)
-
-        // limpando o cache
-        Cache.clean()
-
-        if (conexao.error.length) {
-            res.send({status: false, msg: 'Erro ao conectar com o banco de dados!'})
-        }
 
         try {
             const adminNome     = (!!app.settings.params.nome)  ? app.settings.params.nome  : 'Administrador '+config.sistema
@@ -61,65 +52,96 @@ module.exports = app => {
                 throw new Error('04 - '+__('Os parâmetros %email% e %senha% são obrigatórios para a instalação!'))
             }
 
+            // criando a config geral
+            config.codigo_sistema = 1
+            config.salt         = geraToken()
+            const fileSystem    = require('fs')
+            let configText      = ""
+            configText          += "/**\n"
+            configText          += " * Configurações gerais da API\n"
+            configText          += " *\n"
+            configText          += " * @author  "+adminNome+"\n"
+            configText          += " * @date    "+maskBr(new Date().toLocaleString())+"\n"
+            configText          += " */\n"
+            configText          += "'use strict'\n"
+            configText          += "/**\n"
+            configText          += " * Mantém as configurações da API\n"
+            configText          += " */\n"
+            configText          += "module.exports = "+JSON.stringify(config, null, 2)
+            await fileSystem.writeFile('./config/config.js', configText, function(err) {
+                if (err) {
+                    throw new Error('05 - '+err)
+                }
+            })
+
+            // criando a conexão
+            const conexao = await require(CORE + '/orm/database/driver/'+driver)(configDb)
+            if (conexao.error.length) {
+                throw new Error('06 - '+__('Erro ao conectar com o banco de dados!'))
+            }
+
             // excluindo tudo antes
             if (dropAll && configEnv.debug) {
                 if (! await conexao.query("DROP TABLE IF EXISTS vinculacoes") ) {
-                    throw new Error('05 - '+__('Erro ao tentar excluir a tablela %vinculacoes%!'))
+                    throw new Error('07 - '+__('Erro ao tentar excluir a tablela %vinculacoes%!'))
                 }
                 if (! await conexao.query("DROP TABLE IF EXISTS perfis_rotas") ) {
-                    throw new Error('06 - '+__('Erro ao tentar excluira a tablela %perfis_rotas%!'))
+                    throw new Error('08 - '+__('Erro ao tentar excluira a tablela %perfis_rotas%!'))
                 }
                 if (! await conexao.query("DROP TABLE IF EXISTS rotas") ) {
-                    throw new Error('07 - '+__('Erro ao tentar excluira a tablela %rotas%!'))
+                    throw new Error('09 - '+__('Erro ao tentar excluira a tablela %rotas%!'))
                 }
                 if (! await conexao.query("DROP TABLE IF EXISTS auditorias") ) {
-                    throw new Error('08 - '+__('Erro ao tentar excluira a tablea %auditorias%!'))
+                    throw new Error('10 - '+__('Erro ao tentar excluira a tablea %auditorias%!'))
                 }
                 if (! await conexao.query("DROP TABLE IF EXISTS perfis") ) {
-                    throw new Error('09 - '+__('Erro ao tentar excluira a tablea %perfis%!'))
+                    throw new Error('11 - '+__('Erro ao tentar excluira a tablea %perfis%!'))
                 }
                 if (! await conexao.query("DROP TABLE IF EXISTS unidades") ) {
-                    throw new Error('10 - '+__('Erro ao tentar excluira a tablea %unidades%!'))
+                    throw new Error('12 - '+__('Erro ao tentar excluira a tablea %unidades%!'))
                 }
                 if (! await conexao.query("DROP TABLE IF EXISTS aplicacoes") ) {
-                    throw new Error('11 - '+__('Erro ao tentar excluira a tablela %aplicacoes%!'))
+                    throw new Error('13 - '+__('Erro ao tentar excluira a tablela %aplicacoes%!'))
                 }
                 if (! await conexao.query("DROP TABLE IF EXISTS usuarios") ) {
-                    throw new Error('12 - '+__('Erro ao tentar excluira a tablea %usuarios%!'))
+                    throw new Error('14 - '+__('Erro ao tentar excluira a tablea %usuarios%!'))
                 }
                 if (! await conexao.query("DROP TABLE IF EXISTS municipios") ) {
-                    throw new Error('13 - '+__('Erro ao tentar excluira a tablea %municipios%!'))
+                    throw new Error('15 - '+__('Erro ao tentar excluira a tablea %municipios%!'))
                 }
             }
+
+            // limpando o cache
+            const Cache     = require(CORE + '/lib/cache.js')
+            Cache.clean()
 
             // verificando se existem mais usuários
             const listTables = await conexao.listTables()
             if (listTables.length > 1) {
-                throw new Error('14 - '+__('A instalação inicial já foi inicializada!'))
+                throw new Error('16 - '+__('A instalação inicial já foi inicializada!'))
             }
 
             // instalando as aplicações
             const Aplicacoes = await getTable('mac.aplicacoes')
             if (! await Aplicacoes.createTable({delete:true})) {
-                throw new Error ('15 - ' + Aplicacoes.error)
+                throw new Error ('17 - ' + Aplicacoes.error)
             }
-            let dataAplicacoes = {'ApliId': 1, 'ApliNome': configure('sistema')}
+            let dataAplicacoes = {'ApliId': config.codigo_sistema, 'ApliNome': config.sistema}
             Aplicacoes.forceCreate = true
             if (! await Aplicacoes.save(dataAplicacoes)) {
-                throw new Error('16 - ' + Aplicacoes.error)
+                throw new Error('18 - ' + Aplicacoes.error)
             }
-            config.codigoSistema = 1
 
             // instalando os perfis
             const Perfis = await getTable('mac.perfis')
             if (! await Perfis.createTable({delete: true})) {
-                throw new Error('17 - '+Perfis.error)
+                throw new Error('19 - '+Perfis.error)
             }
 
             // instalando as unidades
             const Unidades = await getTable('mac.unidades')
             if (! await Unidades.createTable({delete: true})) {
-                throw new Error('19 - '+Unidades.error)
+                throw new Error('20 - '+Unidades.error)
             }
             let dataUnidades = {
                 0: {UnidId: 10, UnidCpfCnpj: 10, UnidNome: 'UNIDADE GERAL'},
@@ -128,14 +150,14 @@ module.exports = app => {
             }
             Unidades.forceCreate = true
             if (! await Unidades.saveAll(dataUnidades)) {
-                throw new Error('20 - '+Unidades.error)
+                throw new Error('21 - '+Unidades.error)
             }
 
             // instalando os municípios
             const Municipios = await getTable('mac.municipios')
             Municipios.forceCreate = true
             if (! await Municipios.createTable({delete: true})) {
-                throw new Error('21 - '+Municipios.error)
+                throw new Error('22 - '+Municipios.error)
             }
             // importando municipios
             const linhasMunicipios  = fs.readFileSync(MODULES + '/mac/config/schema/municipios_'+csvMunicipio+'.csv', 'utf8').toString().split('\n')
@@ -159,7 +181,7 @@ module.exports = app => {
                 csvLoop++
                 if (csvLoop >= 1000 || i >= (csvTotal-1)) {
                     if (! await Municipios.saveAll(dataMunicipios)) {
-                        throw new Error('22 - '+Municipios.error)
+                        throw new Error('23 - '+Municipios.error)
                     }
                     csvLoop         = 0
                     dataMunicipios  = []
@@ -169,41 +191,41 @@ module.exports = app => {
             // instalando as aplicações
             const Rotas = await getTable('mac.rotas')
             if (! await Rotas.createTable({delete:true})) {
-                throw new Error ('23 - '+Rotas.error)
+                throw new Error ('24 - '+Rotas.error)
             }
             let dataRotas = {
-                0: {RotaRota: '/listar', RotaAplicacaoId: 1},
-                1: {RotaRota: '/salvar', RotaAplicacaoId: 1},
-                2: {RotaRota: '/excluir', RotaAplicacaoId: 1},
-                3: {RotaRota: '/rotas', RotaAplicacaoId: 1},
-                4: {RotaRota: '/fake', RotaAplicacaoId: 1},
-                5: {RotaRota: '/cadastros', RotaAplicacaoId: 1},
-                6: {RotaRota: '/info', RotaAplicacaoId: 1},
-                7: {RotaRota: '/instalacao', RotaAplicacaoId: 1},
-                8: {RotaRota: '/meu_token', RotaAplicacaoId: 1},
-                9: {RotaRota: '/nova_senha', RotaAplicacaoId: 1},
-                10: {RotaRota: '/novo_token', RotaAplicacaoId: 1},
+                0: {RotaRota: '/listar'},
+                1: {RotaRota: '/salvar'},
+                2: {RotaRota: '/excluir'},
+                3: {RotaRota: '/rotas'},
+                4: {RotaRota: '/fake'},
+                5: {RotaRota: '/cadastros'},
+                6: {RotaRota: '/info'},
+                7: {RotaRota: '/instalacao'},
+                8: {RotaRota: '/meu_token'},
+                9: {RotaRota: '/nova_senha'},
+                10: {RotaRota: '/novo_token'},
             }
             Rotas.forceCreate = true
             if (! await Rotas.saveAll(dataRotas)) {
-                throw new Error('24 - '+Rotas.error)
+                throw new Error('25 - '+Rotas.error)
             }
 
             // salvando os perfis
             let dataPerfis = {
-                0: {PerfId: 1, PerfNome: 'ADMINISTRADOR', PerfAplicacaoId: 1},
-                1: {PerfId: 2, PerfNome: 'DESENVOLVEDOR', PerfAplicacaoId: 1},
-                2: {PerfId: 3, PerfNome: 'SUPERVISOR', PerfAplicacaoId: 1}
+                0: {PerfId: 1, PerfNome: 'ADMINISTRADOR'},
+                1: {PerfId: 2, PerfNome: 'DESENVOLVEDOR'},
+                2: {PerfId: 3, PerfNome: 'SUPERVISOR'}
             }
             Perfis.forceCreate = true
             if (! await Perfis.saveAll(dataPerfis)) {
-                throw new Error('25 - '+Perfis.error)
+                throw new Error('26 - '+Perfis.error)
             }
 
             // instalando usuarios
             const Usuarios = await getTable('mac.usuarios')
             if (! await Usuarios.createTable({delete: true})) {
-                throw new Error('26 - '+Usuarios.error)
+                throw new Error('27 - '+Usuarios.error)
             }
             let dataUsuarios = {
                 0: {
@@ -214,14 +236,14 @@ module.exports = app => {
             }
             Usuarios.forceCreate = true
             if (! await Usuarios.saveAll(dataUsuarios)) {
-                throw new Error('27 - '+Usuarios.error)
+                throw new Error('28 - '+Usuarios.error)
             }
 
             Usuarios.schema.token.hidden = false
             Usuarios.db.mask = false
             const dataUsuario = await Usuarios.find({'type': 'first', 'where':{'Usua.id':1}})
             if (!dataUsuario) {
-                throw new Error('28 - ' + __('Erro ao tentar recuperar dados do usuário!'))
+                throw new Error('29 - ' + __('Erro ao tentar recuperar dados do usuário!'))
             }
 
             // criando a vinculação do usuário admin
@@ -230,39 +252,18 @@ module.exports = app => {
             }
             Perfis.forceCreate = false
             if (! await Perfis.saveAll(dataPerfis)) {
-                throw new Error('29 - '+Perfis.error)
+                throw new Error('30 - '+Perfis.error)
             }
 
             // instalando a auditoria
             const Auditorias = await getTable('mac.auditorias')
             if (! await Auditorias.createTable({delete:true})) {
-                throw new Error ('30 - ' + Auditorias.error)
+                throw new Error ('31 - ' + Auditorias.error)
             }
             global.USUARIO = {id: 1}
             if (! await Auditorias.auditar('Instalação da API realizada com sucesso.', 'instalacao')) {
-                throw new Error ('31 - ' + Auditorias.error)
+                throw new Error ('32 - ' + Auditorias.error)
             }
-
-            // criando a config geral
-            config.salt         = geraToken()
-            const fileSystem    = require('fs')
-            let configText      = ""
-            configText          += "/**\n"
-            configText          += " * Configurações gerais da API\n"
-            configText          += " *\n"
-            configText          += " * @author  "+adminNome+"\n"
-            configText          += " * @date    "+maskBr(new Date().toLocaleString())+"\n"
-            configText          += " */\n"
-            configText          += "'use strict'\n"
-            configText          += "/**\n"
-            configText          += " * Mantém as configurações da API\n"
-            configText          += " */\n"
-            configText          += "module.exports = "+JSON.stringify(config, null, 2)
-            fileSystem.writeFile('./config/config.js', configText, function(err) {
-                if (err) {
-                    throw new Error('32 - '+err)
-                }
-            })
 
             // pós-install
             const fileSqlInstall = './modules/mac/config/schema/install_'+Auditorias.driver+'.sql'
